make memory loading lazy; better support for implicit stores
This commit makes memory loading lazy. As a result, 'buffers' won't be
allocated anywhere till they are used by the program. This doesn't change the
correctness of the simulations themselves but should make them run faster.

In addition, this commit also adds a more direct way of specifying stores that
have been 'implicitly' zerod out. As a part of this effort, the commit adds an
exception for SegFault (and related tests for it).